% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supFTSVD.R
\name{cv_rank_supFTSVD}
\alias{cv_rank_supFTSVD}
\title{Select a rank for supFTSVD by a modified version of cross-validation approach
in Lock and Li (2018)}
\usage{
cv_rank_supFTSVD(
  datlist,
  response,
  interval = NULL,
  ranks = c(1:5),
  resolution = 100,
  CVPhi = FALSE,
  K = 5,
  cvT = 5,
  smooth = 1e-08,
  maxiter = 20,
  epsilon = 1e-04,
  KInd = NULL,
  rsvd_seed = 100,
  conv_criteria = "par",
  rank_fold = 5,
  rfolds_seed = 50,
  rc_thresh = 0.05,
  stratum = NULL
)
}
\arguments{
\item{datlist}{a list with n elements, each is a matrix of dimension p+1 
times m_i. The first row represents time points where measurements were 
obtained. Next p rows represent the function values observed at the time 
points}

\item{response}{a matrix of dimension n by q contains the design matrix for 
supervision of the subject loadings}

\item{interval}{represents the domain of the function. The default is NULL.}

\item{ranks}{a vector of ranks on which grid search will be made}

\item{resolution}{grid size on which the singular functions will be estimated}

\item{CVPhi}{a logical scalar representing whether cross-validation (CV) will be 
performed for determining the smoothness parameters involved with singular 
functions}

\item{K}{number of folds for the CV step}

\item{cvT}{number of initial iterations at which CV will be performed}

\item{smooth}{Smoothing parameter for RKHS norm when CVPhi is FALSE. With 
CVPhi=TRUE, a vector of numeric values for smoothness parameters for grid
search}

\item{maxiter}{Maximum number of iteration. Default: 20.}

\item{epsilon}{Convergence threshold for objective function value}

\item{KInd}{is a vector indices to construct the folds for CV. The default is
NULL that indicates that folds will be constructed randomly inside}

\item{rsvd_seed}{supFTSVD uses random svd for computational advantage while
setting inial values for feature loading vector. This seed number will be used
before setting the initial values. It confirms the reproducibility of your results.}

\item{conv_criteria}{the default is par which sets the threshold on the 
difference between updated and current estimate at M step to stop the iteration. 
It is computationally expensive; thus we recommend to use "cond_lik" instead of "par". 
The result is not sensitive.}

\item{rank_fold}{number of folds to be used in the cross-validation}

\item{rfolds_seed}{seed number to be used in random splitting of the data}

\item{rc_thresh}{threshold on test-data likelihood for rank selection}
}
\value{
a list with following objects:
\itemize{
 \item cv_res : detailed results for different folds
 \item opt_r : optimal rank
}
}
\description{
Select a rank for supFTSVD by a modified version of cross-validation approach
in Lock and Li (2018)
}
\examples{
library(rsvd)
# feature dimension 

pdim<-c(500)

# No of rank-1 components
lr<-2
r<-lr

# Basis dimension
lmd_val<-c(120,80) #lmd_min*(r:1)

# weight for reminder
sig<-0

# noise variance in tensor
Tau2<-1

# Parameters for supervised component
Eta2<-c(1,1.5) # error variance
gam<-as.matrix(cbind(c(1.5,3),c(2,3.4))) # partial slopes



# Grid of Time points
nres<-51
Time<-seq(0,1,length.out=nres)
#Pht<-PhiF[[1]](Tg)
set.seed(nres)
PhiFunc<-lapply(1:r, function(x){
  function(s){
    a<-t(sapply(s,function(j){
      c(1,sapply(2:10, function(i){sqrt(2)*cos((i-1)*(pi)*j)}))
    }))\%*\%matrix(sapply(1:10,function(j){runif(1,-1/j,1/j)}))
    a/norm(a,type="2")
  }
})
PhiF<-sapply(1:r,function(r){sqrt(length(Time))*PhiFunc[[r]](Time)})

set.seed(pdim)
Bval<<-sapply(1:r, function(b){runif(pdim)})
bval<<-Bval*outer(rep(1,pdim),1/apply(Bval,2,norm,type="2"))


# Grid scenario

Mi<-3:8

n<-100
m_i<-sample(Mi,n,replace = TRUE)
set.seed(n)
Vmat<<-cbind(round(runif(n),2),round(rbeta(n,1,1),2))


data<-data_gen_supFTSVD(m_i=m_i,Xmatrix = Vmat,Beta = gam,Xi = bval,
                        PsiF = PhiF,sing_val = lmd_val,SubE_Var = Eta2,
                        Data_Var = Tau2,Tgrid = Time)


# Reproducibility because of the use of random SVD is controlled by rsvd_seed
cv_fit_supF<-cv_rank_supFTSVD(datlist = data$data,
                   response = Vmat,
                   interval = NULL,
                   ranks = c(1:10),
                   resolution = 50,
                   CVPhi = FALSE,
                   K = 5,
                   cvT = 2,
                   smooth = 0.1,
                   maxiter = 100,
                   epsilon = 1e-6,
                   KInd = NULL,rsvd_seed = 25,conv_criteria = "cond_lik")

cv_fit_supF$opt_r
}
