% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supFTSVD.R
\name{ftsvd}
\alias{ftsvd}
\title{Functional tensor singular value decomposition (FTSVD)}
\usage{
ftsvd(
  datlist,
  interval = NULL,
  r = 3,
  resolution = 251,
  CVPhi = FALSE,
  K = 5,
  cvT = 5,
  smooth = 1e-08,
  maxiter = 20,
  epsilon = 1e-04,
  KInd = NULL
)
}
\arguments{
\item{datlist}{a list with n elements, each is a matrix of dimension p+1 
times m_i. The first row represents time points where measurements were 
obtained. Next p rows represent the function values observed at the time 
points}

\item{interval}{represents the domain of the function. The default is NULL.}

\item{r}{number of low-rank components involved in the assumed model}

\item{resolution}{grid size on which the singular functions will be estimated}

\item{CVPhi}{a logical scalar representing whether cross-validation (CV) will be 
performed for determining the smoothness parameters involved with singular 
functions}

\item{K}{number of folds for the CV step}

\item{cvT}{number of initial iterations at which CV will be performed}

\item{smooth}{Smoothing parameter for RKHS norm when CVPhi is FALSE. With 
CVPhi=TRUE, a vector of numeric values for smoothness parameters for grid
search}

\item{maxiter}{Maximum number of iteration. Default: 20.}

\item{epsilon}{Convergence threshold for objective function value}

\item{KInd}{is a vector indices to construct the folds for CV. The default is
NULL that indicates that folds will be constructed randomly inside}
}
\value{
a list with following objects:
\itemize{
 \item A.hat : a matrix of dimension n by r containing the values of 
 conditional means for subjects
 \item B.hat : a matrix of dimension p by r representing the subject loadings
 \item Phi.hat : a matrix of dimension resolution by r representing the 
 estimated singular functions
 \item time : a vector of time points where Phi.hat values are estimated
 \item Lambda : a vector of r lambda values for weights (CP decomposition) of
  low-rank components
 \item r.square : component wise R^2 for tensor model by every rank-1 
 component constructed using conditional mean
 \item accum.r.square : cumulative R^2 values of tensor model by estimated 
 components constructed using conditional means
}
}
\description{
This function models high-dimensional multivariate functional data using the
FTSVD approach (Han et. al. 2023).
}
\examples{
# feature dimension 

pdim<-c(500)

# No of rank-1 components
lr<-2
r<-lr

# Basis dimension
lmd_val<-c(120,80) #lmd_min*(r:1)

# weight for reminder
sig<-0

# noise variance in tensor
Tau2<-1

# Parameters for supervised component
Eta2<-c(1,1.5) # error variance
gam<-as.matrix(cbind(c(1.5,3),c(2,3.4))) # partial slopes



# Grid of Time points
nres<-51
Time<-seq(0,1,length.out=nres)
#Pht<-PhiF[[1]](Tg)
set.seed(nres)
PhiFunc<-lapply(1:r, function(x){
  function(s){
    a<-t(sapply(s,function(j){
      c(1,sapply(2:10, function(i){sqrt(2)*cos((i-1)*(pi)*j)}))
    }))\%*\%matrix(sapply(1:10,function(j){runif(1,-1/j,1/j)}))
    a/norm(a,type="2")
  }
})
PhiF<-sapply(1:r,function(r){sqrt(length(Time))*PhiFunc[[r]](Time)})

set.seed(pdim)
Bval<<-sapply(1:r, function(b){runif(pdim)})
bval<<-Bval*outer(rep(1,pdim),1/apply(Bval,2,norm,type="2"))


# Grid scenario

Mi<-3:8

n<-50
m_i<-sample(Mi,n,replace = TRUE)
set.seed(n)
Vmat<<-cbind(round(runif(n),2),round(rbeta(n,1,1),2))


data<-data_gen_supFTSVD(m_i=m_i,Xmatrix = Vmat,Beta = gam,Xi = bval,
                        PsiF = PhiF,sing_val = lmd_val,SubE_Var = Eta2,
                        Data_Var = Tau2,Tgrid = Time)


set.seed(25) # it is necessary for reproducibility because of the use of random SVD
fit_ftsvd<-ftsvd(datlist = data$data,
                   interval = NULL,
                   r=2,
                   resolution = 50,
                   CVPhi = TRUE,
                   K = 5,
                   cvT = 2,
                   smooth = exp(seq(-2,2,length.out=10)),
                   maxiter = 100,
                   epsilon = 1e-6,
                   KInd = NULL)

}
