% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supFTSVD.R
\name{supFTSVD}
\alias{supFTSVD}
\title{Supervised functional singular value decomposition for high-dimensional 
multivariate functional data.}
\usage{
supFTSVD(
  datlist,
  response,
  interval = NULL,
  r = 3,
  resolution = 100,
  CVPhi = FALSE,
  K = 5,
  cvT = 5,
  smooth = 1e-08,
  maxiter = 20,
  epsilon = 1e-04,
  KInd = NULL
)
}
\arguments{
\item{datlist}{a list with n elements, each is a matrix of dimension p+1 
times m_i. The first row represents time points where measurements were 
obtained. Next p rows represent the function values observed at the time 
points}

\item{response}{a matrix of dimension n by q contains the design matrix for 
supervision of the subject loadings}

\item{interval}{represents the domain of the function. The default is NULL.}

\item{r}{number of low-rank components involved in the assumed model}

\item{resolution}{grid size on which the singular functions will be estimated}

\item{CVPhi}{a logical scalar representing whether cross-validation (CV) will be 
performed for determining the smoothness parameters involved with singular 
functions}

\item{K}{number of folds for the CV step}

\item{cvT}{number of initial iterations at which CV will be performed}

\item{smooth}{Smoothing parameter for RKHS norm when CVPhi is FALSE. With 
CVPhi=TRUE, a vector of numeric values for smoothness parameters for grid
search}

\item{maxiter}{Maximum number of iteration. Default: 20.}

\item{epsilon}{Convergence threshold for objective function value}

\item{KInd}{is a vector indices to construct the folds for CV. The default is
NULL that indicates that folds will be constructed randomly inside}
}
\value{
a list with following objects:
\itemize{
 \item A.hat : a matrix of dimension n by r containing the values of 
 conditional means for subjects
 \item B.hat : a matrix of dimension p by r representing the subject loadings
 \item Phi.hat : a matrix of dimension resolution by r representing the 
 estimated singular functions
 \item time : a vector of time points where Phi.hat values are estimated
 \item Lambda : a vector of r lambda values for weights (CP decomposition) of
  low-rank components
 \item r.square : component wise R^2 for tensor model by every rank-1 
 component constructed using conditional mean
 \item accum.r.square : cumulative R^2 values of tensor model by estimated 
 components constructed using conditional means
 \item r.square.Xb : component wise R^2 for tensor model by every rank-1 
 component constructed using the deterministic part of conditional mean
 \item accum.r.square.Xb : cumulative R^2 values of tensor model by estimated
  components constructed using deterministic parts of conditional means
  \item supR2 : total R^2 achieved by all estimated rank-1 components
  \item supR2.Xb : total R^2 achieved by all estimated rank-1 components 
  involving the deterministic parts of subject loadings
  \item Gamma : a matrix of q by r representing the estimated beta parameters
  \item Xb=Xb : deterministic parts of the conditional means
  \item Sigma2R : r estimated variance parameters involving with subject 
  loading models
  \item Sigma2 : estimated tensor noise variance
  \item cnvrgnt : change of objective function value at different iterations
}
}
\description{
This function performs low-rank decomposition of high-dimensional 
multivariate function following the methodology proposed in Alam (2024+)
}
